requirements.txt
oracledb
langchain
langchain-community
langchain-openai
langchain-experimental
sqlalchemy
python-dotenv
